requirements.txt
boto3
litellm
pandas
datasets
tqdm
python-dotenv
